+++ core-ng/src/main/java/core/framework/api/module/LogConfig.java
@@ -1,6 +1,8 @@
 package core.framework.api.module;
 
+import core.framework.impl.log.ActionLogWriter;
 import core.framework.impl.log.LogForwarder;
+import core.framework.impl.log.TraceLogWriter;
 import core.framework.impl.module.ModuleContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -19,25 +21,41 @@ public final class LogConfig {
         this.context = context;
     }
 
-    public void actionLogPath(Path actionLogPath) {
+    public void writeActionLogToConsole() {
         if (context.test) {
-            logger.info("use console output for action log during test");
+            logger.info("disable action log during test");
         } else {
-            context.logManager.logWriter.actionLogPath(actionLogPath);
+            context.logManager.actionLogWriter = ActionLogWriter.console();
         }
     }
 
-    public void traceLogPath(Path traceLogPath) {
+    public void writeActionLogToFile(Path actionLogPath) {
         if (context.test) {
-            logger.info("use console output for trace log during test");
+            logger.info("disable action log during test");
         } else {
-            context.logManager.logWriter.traceLogPath(traceLogPath);
+            context.logManager.actionLogWriter = ActionLogWriter.file(actionLogPath);
         }
     }
 
-    public void remoteLogHost(String host) {
+    public void writeTraceLogToConsole() {
         if (context.test) {
-            logger.info("disable forwarding log to remote during test");
+            logger.info("disable trace log during test");
+        } else {
+            context.logManager.traceLogWriter = TraceLogWriter.console();
+        }
+    }
+
+    public void writeTraceLogToFile(Path traceLogPath) {
+        if (context.test) {
+            logger.info("disable trace log during test");
+        } else {
+            context.logManager.traceLogWriter = TraceLogWriter.file(traceLogPath);
+        }
+    }
+
+    public void forwardLogToRemote(String host) {
+        if (context.test) {
+            logger.info("disable log forwarding during test");
         } else {
             context.logManager.logForwarder = new LogForwarder(host, context.logManager.appName);
             context.startupHook.add(context.logManager.logForwarder::start);